requirements.txt
appdirs==1.4.3
packaging==16.8
pyparsing==2.2.0
pyserial==3.3
six==1.10.0
appdirs==1.4.3
certifi==2017.4.17
chardet==3.0.3
idna==2.5
packaging==16.8
pyparsing==2.2.0
pyserial==3.3
requests==2.16.5
six==1.10.0
urllib3==1.21.1
